8297539: Consolidate the uses of the int<->float union conversion trick
