+++ src/hotspot/share/runtime/stackValueCollection.cpp
@@ -67,7 +67,7 @@ jdouble StackValueCollection::double_at(int slot) const {
   jint value_array[2] ;
   value_array[0] = at(slot+1)->get_int();
   value_array[1] = at(slot  )->get_int();
-  // 8297539. This matches with Template #8 of cast<To>(From).
+  // This matches with Template #8 of cast<To>(From).
   return *PrimitiveConversions::cast<jdouble *>(value_array);
 #endif
 }
@@ -85,7 +85,7 @@ void StackValueCollection::set_long_at(int slot, jlong value) {
 #else
   // Interpreter stack is reversed in memory:
   // low memory location is in higher java local slot.
-  // 8297539. In arm32 bit builds, no cast<To>(From) can be found by compiler.
+  // In arm32 bit builds, no cast<To>(From) can be found by compiler.
   // So it is not possible to use jint *x_array = PrimitiveConversions::cast<jint*>(&value);
   jint *x_array = (jint*)(&value);
   at(slot+1)->set_int(x_array[0]);
@@ -118,7 +118,7 @@ void StackValueCollection::set_double_at(int slot, jdouble value) {
 #else
   // Interpreter stack is reversed in memory:
   // low memory location is in higher java local slot.
-  // 8297539. This matches with Template #8 of cast<To>(From).
+  // This matches with Template #8 of cast<To>(From).
   // Note: It was expected that this cast can also be used for set_long_at() above.
   jint *x_array = PrimitiveConversions::cast<jint*>(&value);
   at(slot+1)->set_int(x_array[0]);
