+++ src/hotspot/share/utilities/globalDefinitions.hpp
@@ -25,9 +25,9 @@
 #ifndef SHARE_UTILITIES_GLOBALDEFINITIONS_HPP
 #define SHARE_UTILITIES_GLOBALDEFINITIONS_HPP
 
-#include "metaprogramming/primitiveConversions.hpp"
 #include "utilities/compilerWarnings.hpp"
 #include "utilities/debug.hpp"
+#include "utilities/globalJavaValue.hpp"
 #include "utilities/macros.hpp"
 
 // Get constants like JVM_T_CHAR and JVM_SIGNATURE_INT, before pulling in <jvm.h>.
@@ -659,17 +659,6 @@ inline double percent_of(T numerator, T denominator) {
   return denominator != 0 ? (double)numerator / denominator * 100.0 : 0.0;
 }
 
-//----------------------------------------------------------------------------------------------------
-// Special casts
-// Cast floats into same-size integers and vice-versa w/o changing bit-pattern
-
-inline jint    jint_cast    (jfloat  x)  { return PrimitiveConversions::cast<jint>(x);    } // 8297539. Template #5 of cast<To>(From).
-inline jfloat  jfloat_cast  (jint    x)  { return PrimitiveConversions::cast<jfloat>(x);  } // 8297539. Tempalte #5
-
-inline jlong   jlong_cast   (jdouble x)  { return PrimitiveConversions::cast<jlong>(x);   } // 8297539. Tempalte #5
-inline julong  julong_cast  (jdouble x)  { return PrimitiveConversions::cast<julong>(x);  } // 8297539. Tempalte #5
-inline jdouble jdouble_cast (jlong   x)  { return PrimitiveConversions::cast<jdouble>(x); } // 8297539. Tempalte #5
-
 inline jint low (jlong value)                    { return jint(value); }
 inline jint high(jlong value)                    { return jint(value >> 32); }
 
@@ -696,34 +685,6 @@ union jlong_accessor {
 void basic_types_init(); // cannot define here; uses assert
 
 
-// NOTE: replicated in SA in vm/agent/sun/jvm/hotspot/runtime/BasicType.java
-enum BasicType {
-// The values T_BOOLEAN..T_LONG (4..11) are derived from the JVMS.
-  T_BOOLEAN     = JVM_T_BOOLEAN,
-  T_CHAR        = JVM_T_CHAR,
-  T_FLOAT       = JVM_T_FLOAT,
-  T_DOUBLE      = JVM_T_DOUBLE,
-  T_BYTE        = JVM_T_BYTE,
-  T_SHORT       = JVM_T_SHORT,
-  T_INT         = JVM_T_INT,
-  T_LONG        = JVM_T_LONG,
-  // The remaining values are not part of any standard.
-  // T_OBJECT and T_VOID denote two more semantic choices
-  // for method return values.
-  // T_OBJECT and T_ARRAY describe signature syntax.
-  // T_ADDRESS, T_METADATA, T_NARROWOOP, T_NARROWKLASS describe
-  // internal references within the JVM as if they were Java
-  // types in their own right.
-  T_OBJECT      = 12,
-  T_ARRAY       = 13,
-  T_VOID        = 14,
-  T_ADDRESS     = 15,
-  T_NARROWOOP   = 16,
-  T_METADATA    = 17,
-  T_NARROWKLASS = 18,
-  T_CONFLICT    = 19, // for stack value type with conflicting contents
-  T_ILLEGAL     = 99
-};
 
 #define SIGNATURE_TYPES_DO(F, N)                \
     F(JVM_SIGNATURE_BOOLEAN, T_BOOLEAN, N)      \
@@ -872,92 +833,6 @@ inline int type2aelembytes(BasicType t, bool allow_address = false) { return _ty
 inline bool same_type_or_subword_size(BasicType t1, BasicType t2) {
   return (t1 == t2) || (is_subword_type(t1) && type2aelembytes(t1) == type2aelembytes(t2));
 }
-
-// JavaValue serves as a container for arbitrary Java values.
-
-class JavaValue {
-
- public:
-  // 8297539. Define it large enough to hold all possible primitive types.
-  typedef long long JavaCallValue;
-
- private:
-  BasicType _type;
-  JavaCallValue _value;
-
- public:
-  JavaValue(BasicType t = T_ILLEGAL) { _type = t; }
-
-  JavaValue(jfloat value) {
-    _type    = T_FLOAT;
-    // 8297539. This matches with Template #6 of cast<To>(From).
-    _value = PrimitiveConversions::cast<JavaCallValue>(value);
-  }
-
-  JavaValue(jdouble value) {
-    _type    = T_DOUBLE;
-    // 8297539. This matches with Template #5 of cast<To>(From).
-    _value = PrimitiveConversions::cast<JavaCallValue>(value);
-  }
-
- jfloat get_jfloat() const    { return PrimitiveConversions::cast<jfloat>(_value);  } // 8297539. Tempalte #6.
- jdouble get_jdouble() const  { return PrimitiveConversions::cast<jdouble>(_value); } // 8297539. Tempalte #5.
- jint get_jint() const        { return PrimitiveConversions::cast<jint>(_value);    } // 8297539. Tempalte #7.
- jlong get_jlong() const      { return PrimitiveConversions::cast<jlong>(_value);   } // 8297539. Tempalte #1.
- jobject get_jobject() const {
-  #ifdef ARM32
-    // 8297539. In arm32 archs, this call compiles to cast<jobject>(const JavaCallValue&) and
-    // does not match with any of the cast<To>(From) instances.
-    return *(jobject*)(&_value);
-  #else
-    return PrimitiveConversions::cast<jobject>(_value);
-  #endif
- }
- oopDesc* get_oop() const     {
-  #ifdef ARM32
-    // 8297539. In arm32 archs, this call compiles to cast<oopDesc*>(const JavaCallValue&) and
-    // does not match with any of the cast<To>(From) instances.
-    return (oopDesc*)(&_value);
-  #else
-    // 8297539. This matches with Template #4 of cast<To>(From).
-    return PrimitiveConversions::cast<oopDesc*>(_value);
-  #endif
- }
-
- JavaCallValue* get_value_addr() { return &_value; }
- BasicType get_type() const { return _type; }
-
- void set_jfloat(jfloat f)   { _value = PrimitiveConversions::cast<JavaCallValue>(f); } // 8297539. Tempalte #6.
- void set_jdouble(jdouble d) { _value = PrimitiveConversions::cast<JavaCallValue>(d); } // 8297539. Tempalte #5.
- void set_jint(jint i)       { _value = PrimitiveConversions::cast<JavaCallValue>(i); } // 8297539. Tempalte #7.
- void set_jlong(jlong l)     { _value = PrimitiveConversions::cast<JavaCallValue>(l); } // 8297539. Tempalte #1.
- void set_jobject(jobject h) {
-  #ifdef ARM32
-    // 8297539. In arm32 archs, this call compiles to cast<JavaCallValue>(_jobject*&) and
-    // does not match with any of the cast<To>(From) instances.
-    _value = *(JavaCallValue*)h;
-  #else
-    _value = PrimitiveConversions::cast<JavaCallValue>(h);
-  #endif
- }
- void set_oop(oopDesc* o)    {
-  #ifdef ARM32
-    // 8297539. In arm32 archs, this call compiles to cast<JavaCallValue>(oopDesc*&) and
-    // does not match with any of the cast<To>(From) instances.
-    _value = *(JavaCallValue*)o;
-  #else
-    _value = PrimitiveConversions::cast<JavaCallValue>(o);
-  #endif
- }
- void set_type(BasicType t) { _type = t; }
-
- jboolean get_jboolean() const { return PrimitiveConversions::cast<jboolean>(PrimitiveConversions::cast<jint>(_value)); } // 8297539. Tempalte #7.
- jbyte get_jbyte() const       { return PrimitiveConversions::cast<jbyte>(PrimitiveConversions::cast<jint>(_value));    } // 8297539. Tempalte #7.
- jchar get_jchar() const       { return PrimitiveConversions::cast<jchar>(PrimitiveConversions::cast<jint>(_value));    } // 8297539. Tempalte #7.
- jshort get_jshort() const     { return PrimitiveConversions::cast<jshort>(PrimitiveConversions::cast<jint>(_value));   } // 8297539. Tempalte #7.
-
-};
-
 
 // TosState describes the top-of-stack state before and after the execution of
 // a bytecode or method. The top-of-stack value may be cached in one or more CPU
